+++ gantry/commands/__init__.py
@@ -3,8 +3,9 @@ from .completion import completion
 from .config import config
 from .follow import follow
 from .list import list_cmd
+from .logs import logs
 from .main import main
 from .run import run
 from .stop import stop
 
-__all__ = ["main", "follow", "run", "config", "cluster", "stop", "list_cmd", "completion"]
+__all__ = ["main", "follow", "run", "config", "cluster", "stop", "list_cmd", "completion", "logs"]