+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `gantry logs` command.
+
 ## [v1.7.1](https://github.com/allenai/beaker-gantry/releases/tag/v1.7.1) - 2024-07-10
 
 ### Fixed